requirements.txt
apache-airflow==2.3.3
apache-airflow-providers-google==8.4.0
apache-airflow-providers-mysql==3.2.1